Refactor fencis objects to new file
